Only 'setContent' from watch if content differs
Fixed the watch function to prevent the cursor from jumping to the start of the editor content every time the content is changed. See issue #4

Also made id property optional with a unique id set if it doesn't exist
And made content property redundant by pulling the value into the setContent on init. I have left it in so that current installs don't break but it isn't really needed anymore